Move some functions into a new library libcfcrypto
A new library for crypto-related things that have nothing to do
with policy and thus don't need to be part of libpromises. This
should allow us to stop linking to libpromises in some of our
binaries and libraries.

For start, let's move the initialization functions.
