+++ libcfcrypto/Makefile.am
@@ -21,29 +21,18 @@
 # (COSL) may apply to this file if you as a licensee so wish it. See
 # included file COSL.txt.
 #
-noinst_LTLIBRARIES = libcfnet.la
 
-AM_CPPFLAGS = -I$(top_srcdir)/libntech/libutils \
-	-I$(top_srcdir)/libcfcrypto \
-	-I$(top_srcdir)/libpromises \
-	$(PCRE_CPPFLAGS) \
-	$(OPENSSL_CPPFLAGS)
+projlib_LTLIBRARIES = libcfcrypto.la
 
-libcfnet_la_SOURCES = \
-	addr_lib.c addr_lib.h \
-	client_protocol.c client_protocol.h cfnet.h\
-	client_code.c client_code.h \
-	classic.c classic.h \
-	communication.c communication.h \
-	connection_info.c connection_info.h \
-	conn_cache.c conn_cache.h \
-	key.c key.h \
-	misc.c \
-	net.c net.h \
-	policy_server.c policy_server.h \
-	protocol.c protocol.h \
-	protocol_version.c protocol_version.h \
-	server_code.c server_code.h \
-	stat_cache.c stat_cache.h \
-	tls_client.c tls_client.h \
-	tls_generic.c tls_generic.h
+AM_LDFLAGS = $(CORE_LDFLAGS) $(OPENSSL_LDFLAGS)
+
+AM_CPPFLAGS = \
+	-I$(srcdir)/../libntech/libutils \
+	-I$(srcdir)/../libpromises \
+	$(CORE_CPPFLAGS) $(OPENSSL_CPPFLAGS)
+
+AM_CFLAGS = $(CORE_CFLAGS) $(OPENSSL_CPPFLAGS)
+
+libcfcrypto_la_LIBS = $(srcdir)/../libntech/libutils.la $(OPENSSL_LIBS)
+
+libcfcrypto_la_SOURCES = crypto_init.c crypto_init.h
